+++ packages/jest-snapshot/src/printSnapshot.ts
@@ -15,6 +15,7 @@ import {
   DIFF_EQUAL,
   DIFF_INSERT,
   Diff,
+  DiffOptionsColor,
   diffLinesUnified,
   diffStringsRaw,
   diffStringsUnified,
@@ -30,8 +31,60 @@ import {
   matcherHint,
 } from 'jest-matcher-utils';
 import prettyFormat = require('pretty-format');
+
+import {
+  aBackground2,
+  aBackground3,
+  aForeground2,
+  aForeground3,
+  bBackground2,
+  bBackground3,
+  bForeground2,
+  bForeground3,
+} from './colors';
 import {MatchSnapshotConfig} from './types';
 import {deserializeString, minify, serialize} from './utils';
+
+type Chalk = chalk.Chalk;
+
+export const getSnapshotColorForChalkInstance = (
+  chalkInstance: Chalk,
+): DiffOptionsColor => {
+  const level = chalkInstance.level;
+
+  if (level === 3) {
+    return chalkInstance
+      .rgb(aForeground3[0], aForeground3[1], aForeground3[2])
+      .bgRgb(aBackground3[0], aBackground3[1], aBackground3[2]);
+  }
+
+  if (level === 2) {
+    return chalkInstance.ansi256(aForeground2).bgAnsi256(aBackground2);
+  }
+
+  return chalkInstance.magenta.bgYellowBright;
+};
+
+export const getReceivedColorForChalkInstance = (
+  chalkInstance: Chalk,
+): DiffOptionsColor => {
+  const level = chalkInstance.level;
+
+  if (level === 3) {
+    return chalkInstance
+      .rgb(bForeground3[0], bForeground3[1], bForeground3[2])
+      .bgRgb(bBackground3[0], bBackground3[1], bBackground3[2]);
+  }
+
+  if (level === 2) {
+    return chalkInstance.ansi256(bForeground2).bgAnsi256(bBackground2);
+  }
+
+  return chalkInstance.cyan.bgWhiteBright; // also known as teal
+};
+
+export const aSnapshotColor = getSnapshotColorForChalkInstance(chalk);
+export const bReceivedColor = getReceivedColorForChalkInstance(chalk);
 
 export const noColor = (string: string) => string;
 
@@ -50,6 +103,9 @@ export const matcherHintFromConfig = (
   isUpdatable: boolean,
 ): string => {
   const options: MatcherHintOptions = {isNot, promise};
+  if (isUpdatable) {
+    options.receivedColor = bReceivedColor;
+  }
 
   let expectedArgument = '';
 
@@ -64,7 +120,9 @@ export const matcherHintFromConfig = (
       options.secondArgumentColor = BOLD_WEIGHT;
     } else if (typeof inlineSnapshot === 'string') {
       options.secondArgument = SNAPSHOT_ARG;
-      if (!isUpdatable) {
+      if (isUpdatable) {
+        options.secondArgumentColor = aSnapshotColor;
+      } else {
         options.secondArgumentColor = noColor;
       }
     }
@@ -74,8 +132,8 @@ export const matcherHintFromConfig = (
       options.expectedColor = BOLD_WEIGHT;
     } else if (typeof inlineSnapshot === 'string') {
       expectedArgument = SNAPSHOT_ARG;
-      if (!isUpdatable) {
-        options.expectedColor = noColor;
+      if (isUpdatable) {
+        options.expectedColor = aSnapshotColor;
       }
     }
   }
@@ -184,14 +242,14 @@ export const printSnapshotAndReceived = (
 ): string => {
   const aAnnotation = 'Snapshot';
   const bAnnotation = 'Received';
-  const aColor = EXPECTED_COLOR;
-  const bColor = RECEIVED_COLOR;
+  const aColor = aSnapshotColor;
+  const bColor = bReceivedColor;
   const options = {
     aAnnotation,
     aColor,
     bAnnotation,
     bColor,
-    changeLineTrailingSpaceColor: chalk.bgYellow,
+    changeLineTrailingSpaceColor: noColor,
     commonLineTrailingSpaceColor: chalk.bgYellow,
     emptyFirstOrLastLinePlaceholder: '↵', // U+21B5
     expand,